requirements.txt
fastapi==0.99.2
uvicorn[standard]==0.23.2
sqlmodel==0.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.24.1
pytest==7.4.0
pytest-asyncio==0.22.0
alembic==1.11.1
python-multipart==0.0.6
